@@ -1,56 +1,75 @@
 {
-	"name": "@egjs/<%= componentName.toLowerCase() %>",
-	"version": "<%= options.version %>",
-	"main": "dist/<%= componentName.toLowerCase() %>.js",
-	"scripts": {
+  "name": "@egjs/<%= componentname %>",
+  "version": "<%= version %>",
+  "main": "dist/<%= componentname %>.js",
+  "es2015": "src/<%= ComponentName %>.js",
+  "scripts": {
     "start": "webpack-dev-server --open",
-    "build": "webpack --env=development",
-    "build:prod": "webpack --env=production && webpack --env=production-packaged",
+    "build": "webpack --env production && webpack --env packaged",
+    "test": "karma start",
+    "test:chrome": "karma start --chrome",
     "lint": "eslint src",
     "jsdoc": "rm -rf ./doc && jsdoc -c jsdoc.json",
+    "coverage": "karma start --coverage",
+    "demo:start": "npm run demo:prebuild-latest && npm run demo:updateVersion && jekyll serve -s demo",
+    "demo:build": "npm run demo:updateVersion && jekyll build -s demo",
+    "demo:prebuild-version": "cpx 'dist/**/*' demo/release/$npm_package_version/dist --clean && cpx 'doc/**/*' demo/release/$npm_package_version/doc --clean",
+    "demo:prebuild-latest": "cpx 'dist/**/*' demo/release/latest/dist --clean && cpx 'doc/**/*' demo/release/latest/doc --clean",
+    "demo:deploy": "npm run build && npm run jsdoc && npm run demo:prebuild-version && npm run demo:prebuild-latest && npm run demo:build && gh-pages -d demo/_site",
+    "demo:setup": "cpx 'node_modules/@egjs/common-demo/**/*' 'demo/' && rm demo/package.json",    
+    "demo:updateVersion": "echo $npm_package_version > demo/_data/version.yml", 
     "prepush": "npm run lint",
-    "commitmsg": "node config/validate-commit-msg.js",
-    "test": "karma start"
-	},
-	"author": "<%= options.author %>",
-	"license": "<%= options.license %>",<% if(options.extendsComponent){ %>
-	"dependencies": {
-    "@egjs/component": "^2.0.0-rc.1"
-	},<% } %>
-	"devDependencies": {
-    "babel-core": "^6.24.0",
-    "babel-loader": "^6.4.1",
+    "commitmsg": "node config/validate-commit-msg.js"
+  },
+  "author": {
+    "name": "<%= author %>"
+  },
+  "namespace": {
+    "eg": "eg"
+  },
+  "license": "<%= license %>",
+  <% if(extendsComponent){ %>
+  "dependencies": {
+    "@egjs/component": "^2.0.0"
+  },
+  <% } %>
+  "devDependencies": {
+    "@egjs/common-demo": "github:naver/egjs#common-demo",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.0",
     "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-es2015": "^6.24.0",
-    "chai": "^3.5.0",
+    "babel-plugin-transform-object-assign": "^6.22.0",
+    "babel-preset-es2015": "^6.24.1",
+    "chai": "^4.0.2",
+    "clean-webpack-plugin": "^0.1.16",
     "cpx": "^1.5.0",
-    "egjs-jsdoc-template": "^1.3.1",
-    "eslint": "^3.18.0",
+    "egjs-jsdoc-template": "1.3.1",
+    "eslint": "^3.19.0",
     "eslint-config-naver": "^1.0.0",
-    "eslint-loader": "^1.7.0",
-    "gh-pages": "^0.12.0",
-    "husky": "^0.13.3",
+    "eslint-loader": "^1.8.0",
+    "gh-pages": "^1.0.0",
+    "husky": "^0.14.1",
     "inject-loader": "^3.0.0",
-    "iscroll": "^5.2.0",
-    "jquery": "^1.7.0",
+    "istanbul-instrumenter-loader": "^2.0.0",
     "jsdoc": "^3.4.3",
-    "karma": "^1.5.0",
+    "karma": "^1.7.0",
     "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.3",
-    "karma-phantomjs-launcher": "^1.0.4",
     "karma-sinon": "^1.0.5",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^2.0.3",
-    "mocha": "^3.2.0",
-    "phantomjs-polyfill": "0.0.2",
-    "phantomjs-polyfill-object-assign": "0.0.2",
-    "phantomjs-prebuilt": "^2.1.14",
-    "uglify-js": "^2.8.16",
-    "uglifyjs-webpack-plugin": "^0.3.1",
-    "webpack": "^2.3.2",
-    "webpack-dev-server": "^2.4.2",
+    "lite-fixture": "^1.0.2",
+    "mocha": "^3.4.2",
+    "sinon": "^2.3.5",
+    "string-replace-webpack-plugin": "0.1.3",
+    "uglify-js": "^2.8.28",
+    "uglifyjs-webpack-plugin": "^0.4.6",
+    "webpack": "^2.6.1",
+    "webpack-dev-server": "^2.5.0",
     "webpack-merge": "^4.1.0",
-    "write-file-webpack-plugin": "^4.0.0"
-	}
+    "write-file-webpack-plugin": "^4.1.0"
+  }
 }
