@@ -1,57 +1,61 @@
 module.exports = function(config) {
-	config.set({
-		frameworks: ["mocha", "chai"],
+  var karmaConfig = {
+    frameworks: ['mocha', 'chai', 'sinon'],
 
-		files: [
-			"./node_modules/phantomjs-polyfill/bind-polyfill.js",
-			"./node_modules/phantomjs-polyfill-object-assign/object-assign-polyfill.js",
-			"./test/unit/*.spec.js"
-		],
+    // list of files / patterns to load in the browser
+    files: [
+      './node_modules/lite-fixture/index.js',
+      './test/unit/**/*.spec.js'
+    ],
 
-		preprocessors: {
-			"./test/unit/*.spec.js": ["webpack"]
-		},
+    client: {
+      mocha: {
+        opts: './mocha.opts' 
+      }
+    },
 
-		webpack: {
-			devtool: "inline-source-map",
-			module: {
-				rules: [
-					{
-						test: /\.js$/,
-						exclude: /node_modules/,
-						loader: "babel-loader",
-						options: {
-							presets: ["es2015"],
-							plugins: ["add-module-exports"]
-						}
-					}
-				]
-			}
-		},
+    webpack: {
+      devtool: 'inline-source-map',
+      module: {
+          rules: [
+              {
+                  test: /\.js$/,
+                  exclude: /node_modules/,
+                  loader: "babel-loader",
+              }
+          ]
+      }
+    },
 
-		webpackMiddleware: {
-			noInfo: true
-		},
+    // preprocess matching files before serving them to the browser
+    // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
+    preprocessors: {
+      './test/**/*.spec.js': config.coverage ? ['webpack'] : ['webpack', 'sourcemap']
+    },
 
-		reporters: ["mocha"],
+    browsers: [],
+    
+    reporters: ['mocha'],
+    webpackMiddleware: {
+        noInfo: true
+    }
+  };
+  
+  karmaConfig.browsers.push(config.chrome ? "Chrome" : "ChromeHeadless");
 
-		browsers: ["PhantomJS_custom"],
+  if(config.coverage) {
+    karmaConfig.reporters.push('coverage-istanbul');
+    karmaConfig.coverageIstanbulReporter = {
+      reports: [ 'text-summary', 'html'],
+      dir: './coverage'
+    };
+    karmaConfig.webpack.module.rules.unshift({
+        test: /\.js$/,
+        exclude: /(node_modules|test)/,
+        loader: 'istanbul-instrumenter-loader'
+    });
+    karmaConfig.singleRun = true;
+  }
 
-		customLaunchers: {
-			"PhantomJS_custom": {
-				base: "PhantomJS",
-				options: {
-					viewportSize: {
-						width: 1024,
-						height: 960
-					}
-				}
-			}
-		},
-
-		client: {
-			useIframe: false,
-			runInParent: true
-		}
-	});
+  config.set(karmaConfig);
 };
